move test to ./tests
